requirements.txt
pandas
numpy
matplotlib
pygame
sounddevice
scipy
Pillow
PyQt5

